svg/animations/scripted/end-element-on-inactive-element.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml"><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><rect width="100" height="100" fill="red"><set id="anim" attributeName="fill" begin="indefinite" to="green"/></rect><script>async_test(t =&gt; { let anim = document.getElementById(&quot;anim&quot;); onload = t.step_func(() =&gt; { anim.endElement(); anim.beginElement(); requestAnimationFrame(t.step_func_done(() =&gt; { assert_equals(getComputedStyle(anim.parentNode).fill, &quot;rgb(0, 128, 0)&quot;); })); }); }, &quot;endElement() on an inactive element&quot;);</script></svg>
