LayoutTests/imported/w3c/web-platform-tests/svg/animations/scripted/end-element-on-inactive-element-expected.txt


FAIL endElement() on an inactive element assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
